added: scroll-to and highlight selected text even for pending follow up chip on chatInput, moved: the general logics into services/selectionGlow.ts and load it up for both chatMessageand chatInput
